% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResizeMatrix.r
\name{ResizeMatrix}
\alias{ResizeMatrix}
\title{ResizeMatrix}
\usage{
ResizeMatrix(matrice.mtx, newDim.num = dim(matrice.mtx))
}
\arguments{
\item{matrice.mtx}{<matrix>: a Matrix to resize}

\item{newDim.num}{<integer>: the number of rows and cols in resized matrix.}
}
\value{
resized matrix
}
\description{
Resize a matrix in new dimension
}
\examples{
matrice.mtx <- matrix(0,11,11)
matrice.mtx[which(as.logical(1:(11*11)\%\%2))] <- 1:ceiling((11*11)/2)
matrice.mtx[2,] <- 100
matrice.mtx[,7] <- 200
matrice.mtx
ResizeMatrix(matrice.mtx=matrice.mtx, newDim.num=c(7,7))
ResizeMatrix(matrice.mtx=matrice.mtx, newDim.num=c(7,7))
ResizeMatrix(matrice.mtx=matrice.mtx, newDim.num=c(13,13))
ResizeMatrix(matrice.mtx=matrice.mtx, newDim.num=c(13,13))
}
