% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResizeMatrix.r
\name{ResizeMatrix}
\alias{ResizeMatrix}
\title{ResizeMatrix}
\usage{
ResizeMatrix(
  mat.mtx = NULL,
  rowDim.num = NULL,
  colDim.num = NULL,
  rowKeep.ndx = NULL,
  colKeep.ndx = NULL
)
}
\arguments{
\item{mat.mtx}{<matrix>: a Matrix to resize}

\item{rowDim.num}{<integer>: the number of rows in reiszed matrix}

\item{colDim.num}{<integer>: the number of columns in reiszed matrix}

\item{rowKeep.ndx}{<integer>: index of one row to keep in resized matrix}

\item{colKeep.ndx}{<integer>: index of one column to keep in resized matrix}
}
\value{
resized matrix
}
\description{
Resize a matrix in new dimension
}
\examples{
mat.mtx <- matrix(0,11,11)
mat.mtx[which(as.logical(1:(11*11)\%\%2))] <- 1:ceiling((11*11)/2)
mat.mtx[2,] <- 100
mat.mtx[,7] <- 200
mat.mtx
ResizeMatrix(mat.mtx=mat.mtx, rowDim.num=7, colDim.num=7, rowKeep.ndx=NULL, colKeep.ndx=NULL)
ResizeMatrix(mat.mtx=mat.mtx, rowDim.num=7, colDim.num=7, rowKeep.ndx=2, colKeep.ndx=7)
ResizeMatrix(mat.mtx=mat.mtx, rowDim.num=13, colDim.num=13, rowKeep.ndx=NULL, colKeep.ndx=NULL)
ResizeMatrix(mat.mtx=mat.mtx, rowDim.num=13, colDim.num=13, rowKeep.ndx=2, colKeep.ndx=7)
}
